feat: changes on task-card
